Use default plural forms expression when plurals not translated.
Previously, if a plural were not translated, the current language's
plural forms expression would be used to select from the two msgids
passed into ngettext. When the current language had different plurals
from English, this could result in `undefined` being returned instead of
a string.

This makes the fallbacks use the default plural forms expression so that
a valid string is returned whether translated or not.